requirements.txt
Pillow==7.0.0
pyaes==1.6.1
pyasn1==0.4.5
pytz==2019.1
rsa==4.0
Telethon==1.8.0
